move config files to a config module to make goreleaser happy
