+++ Makefile
@@ -1,7 +1,9 @@
-build: generate
-	go build
+build:
+	CGO_ENABLED=0 go build -ldflags "-s -w"
+
 generate:
-	go generate
+
 clean:
 	rm -f nsd_exporter
+
 all: clean build
