+++ config/metric_config.go
@@ -1,4 +1,4 @@
-package main
+package config
 
 import (
 	"embed"
@@ -8,12 +8,12 @@ import (
 	"strings"
 
 	"github.com/prometheus/client_golang/prometheus"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 // ClassifierCRM have share files for CRM114 classifer
 //
-//go:embed config
+//go:embed config.yaml
 var configFS embed.FS
 
 var stringToValueType = map[string]prometheus.ValueType{
@@ -22,17 +22,17 @@ var stringToValueType = map[string]prometheus.ValueType{
 	"untyped": prometheus.UntypedValue,
 }
 
-type metricConfig struct {
-	Metrics      map[string]*metric      `yaml:"metrics"`
-	LabelMetrics map[string]*labelMetric `yaml:"label_metrics"`
+type MetricConfig struct {
+	Metrics      map[string]*Metric      `yaml:"metrics"`
+	LabelMetrics map[string]*LabelMetric `yaml:"label_metrics"`
 }
 
-type metric struct {
+type Metric struct {
 	Help string               `yaml:"help"`
 	Type prometheus.ValueType `yaml:"type"`
 }
 
-type labelMetric struct {
+type LabelMetric struct {
 	Help   string               `yaml:"help"`
 	Name   string               `yaml:"name"`
 	Labels []string             `yaml:"labels"`
@@ -51,7 +51,7 @@ func stringToPromType(s string) prometheus.ValueType {
 }
 
 // Serialize YAML into correct types
-func (m *metric) UnmarshalYAML(unmarshal func(interface{}) error) error {
+func (m *Metric) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	var tmp struct {
 		Help string
 		Type string
@@ -66,7 +66,7 @@ func (m *metric) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	return nil
 }
 
-func (m *labelMetric) UnmarshalYAML(unmarshal func(interface{}) error) error {
+func (m *LabelMetric) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	var tmp struct {
 		Name   string
 		Help   string
@@ -86,12 +86,12 @@ func (m *labelMetric) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	return nil
 }
 
-func loadConfig(path string, metricConf *metricConfig) error {
+func LoadConfig(path string, metricConf *MetricConfig) error {
 	var b []byte
 	var err error
 
 	if path == "" {
-		b, err = configFS.ReadFile("config/config.yaml")
+		b, err = configFS.ReadFile("config.yaml")
 		if err != nil {
 			return err
 		}
